refactored project: split basepage code into modules and created modules objects in Page Objects. Updated tests to run successfully after refactor
