+++ src/main/java/constants/ConstantsMessages.java
@@ -2,11 +2,11 @@ package constants;
 
 public class ConstantsMessages {
 
-    public static final String myAccountTitle = "My account - My Store";
-    public static final String loginPageH1 = "AUTHENTICATION";
-    public static final String womenPageH1 = "WOMEN";
+    public static final String myAccountTitle = "MY ACCOUNT";
+    public static final String loginPageTitle = "AUTHENTICATION";
+    public static final String womenPageTitle = "WOMEN";
     public static final String validUsername = "Qa Test";
-    public static final String productAddedSuccessfullyH2 = "Product successfully added to your shopping cart";
+    public static final String productAddedSuccessfullyTitle = "Product successfully added to your shopping cart";
     public static final String productNameSKUDemo1 = "Faded Short Sleeve T-shirts";
     public static final String productNameSKUDemo2 = "Blouse";
     public static final String cartItem1Product = "1 Product";
