+++ src/test/java/logintests/PositiveLoginTest.java
@@ -7,6 +7,7 @@ import constants.ConstantsURLs;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import pages.BasePage;
+import pages.HomePage;
 import pages.LoginPage;
 import pages.MyAccountPage;
 
@@ -16,21 +17,23 @@ public class PositiveLoginTest extends BaseTest {
     public void logInTestUsingSignInButton() {
         log.info("Starting Log In test using Sign in button");
 
-        BasePage basePage = new BasePage(driver, log);
+        //Open main URL
         openURL(ConstantsURLs.baseURL);
-        basePage.clickOnSignInLink();
 
-        LoginPage loginPage = new LoginPage(driver, log);
+        //Navigate to Login page and sign in
+        HomePage homePage = new HomePage(driver, log);
+        LoginPage loginPage = homePage.getHeaderModule().clickOnSignInLink();
         MyAccountPage myAccountPage = loginPage.logInUserSignInBtn(ConstantsCredentials.validEmail,
                 ConstantsCredentials.validPassword);
         log.info("User successfully logged in");
 
+        //Check that correct user is signed in and landed on My Account page
         Assert.assertEquals(ConstantsURLs.myAccountURL, myAccountPage.getCurrentUrl());
         Assert.assertEquals(ConstantsMessages.myAccountTitle, myAccountPage.getCurrentPageTitle());
-        log.info("User is on " + myAccountPage.getCurrentPageH1() + " page");
-        Assert.assertEquals(ConstantsMessages.validUsername, basePage.getUsername());
+        log.info("User is on " + myAccountPage.getCurrentPageTitle() + " page");
+        Assert.assertEquals(ConstantsMessages.validUsername, myAccountPage.getHeaderModule().getUsername());
         log.info("Correct username is displayed in header");
-        Assert.assertTrue(basePage.isSignOutLinkVisible());
+        Assert.assertTrue(myAccountPage.getHeaderModule().isSignOutLinkVisible());
         log.info("Sign out button is displayed in header");
     }
 
@@ -38,21 +41,23 @@ public class PositiveLoginTest extends BaseTest {
     public void logInTestUsingEnterKey() {
         log.info("Starting Log In test using Enter key");
 
-        BasePage basePage = new BasePage(driver, log);
+        //Open main URL
         openURL(ConstantsURLs.baseURL);
-        basePage.clickOnSignInLink();
 
-        LoginPage loginPage = new LoginPage(driver, log);
+        //Navigate to Login page and sign in
+        HomePage homePage = new HomePage(driver, log);
+        LoginPage loginPage = homePage.getHeaderModule().clickOnSignInLink();
         MyAccountPage myAccountPage = loginPage.logInUserEnterKey(ConstantsCredentials.validEmail,
                 ConstantsCredentials.validPassword);
         log.info("User successfully logged in");
 
+        //Check that correct user is signed in and landed on My Account page
         Assert.assertEquals(ConstantsURLs.myAccountURL, myAccountPage.getCurrentUrl());
         Assert.assertEquals(ConstantsMessages.myAccountTitle, myAccountPage.getCurrentPageTitle());
-        log.info("User is on " + myAccountPage.getCurrentPageH1() + " page");
-        Assert.assertEquals(ConstantsMessages.validUsername, basePage.getUsername());
+        log.info("User is on " + myAccountPage.getCurrentPageTitle() + " page");
+        Assert.assertEquals(ConstantsMessages.validUsername, myAccountPage.getHeaderModule().getUsername());
         log.info("Correct username is displayed in header");
-        Assert.assertTrue(basePage.isSignOutLinkVisible());
+        Assert.assertTrue(myAccountPage.getHeaderModule().isSignOutLinkVisible());
         log.info("Sign out button is displayed in header");
     }
 }
